Change the version type to use unsigned long
Versions should not be able to go to a negative range

+++ include/kinetic_notation/types.h
@@ -3,11 +3,12 @@
 
 // std
 #include <stdbool.h>
+#include <stddef.h>
 
 typedef struct {
-  int major;
-  int minor;
-  int patch;
+  size_t major;
+  size_t minor;
+  size_t patch;
 } kn_version;
 
 #endif // KINETIC_NOTATION_TYPES_H
